### 2019-12-26 #1985 @Kevin-Mattheus-Moerman
@openjournals/dev the paper is here, on a different branch: https://github.com/castelao/CoTeDe/tree/joss/joss
What is the process for handling this again? Thanks
