### 2019-12-26 #1985 @castelao
Thanks @Kevin-Mattheus-Moerman . I can merge the joss branch into the master if that helps.

A suggestion of reviewers: efiring and koldunovn
