### 2019-12-27 1e5c62a
Updating tests to use .attrs instead of .attributes
