### 2019-12-29 861a3d2
Use os.path.join instead of explicit '/' in config
Although there were no reports on that, it should be a problem for
windows machine. Instead of definiing explicitly the path, use join to
compose the path with the right syntax for the running machine.
