### 2019-12-29 6bf6a18
Convert user config dict into ordered dictionary
All JSON saved configs are read as OrderedDict. If the user provides an
inline config dictionary, converts it into an OrderedDict for
consistency.
