### 2019-12-29 a67e885
Need to open first the user custom config json
