### 2019-12-30 195eddc
Forgot to clean old load_cfg (related to 4d0fe34)
