+++ cotede/utils/utils.py
@@ -78,52 +78,3 @@ def loadPQCCollection_pandas(filename):
     tar = tarfile.open(filename, "r:*")
     tar.extractall(path=tmpdir)
     shutil.rmtree(tmpdir)
-
-
-def load_cfg(cfg=None):
-    """ Load the QC configurations
-
-        The possible inputs are:
-            - None: Will use the CoTeDe's default configuration
-
-            - Config name [string]: A string with the name of a json file
-                describing the QC procedure. It will first search among
-                the build in pre-set (cotede, eurogoos, gtspp or argo),
-                otherwise it will search in ~/.cotederc/cfg
-
-            - User configs [dict]: a dictionary composed by the variables
-                to be evaluated as keys, and inside it another dictionary
-                with the tests to perform as keys. example
-                {'main':{
-                    'valid_datetime': None,
-                    },
-                'temperature':{
-                    'global_range':{
-                        'minval': -2.5,
-                        'maxval': 45,
-                        },
-                    },
-                }
-    """
-    # A given manual configuration has priority
-    if type(cfg) is dict:
-        #self.cfg = cfg
-        #self.logger.debug("%s - User's QC cfg." % self.name)
-        return cfg
-
-    if cfg is None:
-        cfg = 'cotede'
-
-    try:
-        # If cfg is available in qc_cfg, use it
-        cfg = json.loads(pkg_resources.resource_string('cotede',
-            "qc_cfg/%s.json" % cfg))
-        #self.logger.debug("%s - QC cfg: %s" % (self.name, cfg))
-        return cfg
-    except:
-        # Otherwise, try to load from user's directory
-        cfg = json.load(open(
-            os.path.join(cotede_dir(), 'cfg/%s.json' % cfg)))
-        #self.logger.debug("%s - QC cfg: ~/.cotederc/%s" %
-        #            (self.name, cfg))
-        return cfg
