### 2019-12-30 22c83a5
Updating [spike|gradient] depth conditional test
Both tests now use the standard QCCheck class.
