### 2019-12-30 ef5cef3
Updating density inversion test
It now uses the standard QCCheck class.
