### 2019-12-30 cd2b328
Cleanning tests that are based on QCCheck
