### 2019-12-31 2133f75
Updating global range test
It now uses the standard QCCheck class.
