+++ cotede/qctests/global_range.py
@@ -10,27 +10,17 @@
     - properly handle inf, nan?
 """
 
+import logging
+
 import numpy as np
 from numpy import ma
 
+from .qctests import QCCheck
 
-class GlobalRange(object):
-    def __init__(self, data, varname, cfg, autoflag=True):
-        self.data = data
-        self.varname = varname
-        self.cfg = cfg
 
-        self.set_features()
-        if autoflag:
-            self.test()
+module_logger = logging.getLogger(__name__)
 
-    def keys(self):
-        return self.features.keys() + \
-            ["flag_%s" % f for f in self.flags.keys()]
-
-    def set_features(self):
-        self.features = {}
-
+class GlobalRange(QCCheck):
     def test(self):
         self.flags = {}
         assert ('minval' in self.cfg) and ('maxval' in self.cfg), \
@@ -43,27 +33,12 @@ class GlobalRange(object):
         minval = self.cfg['minval']
         maxval = self.cfg['maxval']
 
-        try:
-            flag_good = self.cfg['flag_good']
-        except:
-            flag_good = 1
-        try:
-            flag_bad = self.cfg['flag_bad']
-        except:
-            flag_bad = 4
+        feature = self.data[self.varname]
 
-        flag = np.zeros(self.data[self.varname].shape, dtype='i1')
-
-        idx = self.data[self.varname] < minval
-        flag[np.nonzero(idx)] = flag_bad
-
-        idx = self.data[self.varname] > maxval
-        flag[np.nonzero(idx)] = flag_bad
-
-        idx = (self.data[self.varname] >= minval) \
-                & (self.data[self.varname] <= maxval)
-        flag[np.nonzero(idx)] = flag_good
-
+        flag = np.zeros(feature.shape, dtype='i1')
+        flag[np.nonzero(feature < minval)] = self.flag_bad
+        flag[np.nonzero(feature > maxval)] = self.flag_bad
+        idx = (feature >= minval) & (feature <= maxval)
+        flag[np.nonzero(idx)] = self.flag_good
         flag[ma.getmaskarray(self.data[self.varname])] = 9
-
         self.flags['global_range'] = flag
