### 2019-12-31 c68b55e
Updating profile envelop test
It now uses the standard QCCheck class.
