+++ tests/qctests/test_qc_profile_envelop.py
@@ -5,20 +5,38 @@
 """
 
 import numpy as np
-from numpy import ma
 
-from cotede.qctests import profile_envelop
+from cotede.qctests.profile_envelop import ProfileEnvelop, profile_envelop
 from data import DummyData
 
 
 def test():
     profile = DummyData()
 
-    cfg = [
-        ["> 0", "<= 25", -2, 37],
-        ["> 25", "<= 50", -2, 36]]
+    cfg = [["> 0", "<= 25", -2, 37], ["> 25", "<= 50", -2, 36]]
 
-    flags = profile_envelop(profile, cfg, 'TEMP')
+    flags = profile_envelop(profile, cfg, "TEMP")
 
     # Check for BUG #9.
     assert 9 in flags
+
+
+def test_class():
+    """
+
+       ATTENTION, unrealistic limits used just for testing purposes.
+    """
+    data = DummyData()
+    cfg = [
+        ["> 0", "<= 149", -2, 37],
+        ["> 150", "<= 999", -2, 33],
+        ["> 999", "<= 12000", -1.5, 4],
+    ]
+    y = ProfileEnvelop(data, varname="TEMP", cfg=cfg)
+
+    assert hasattr(y, "features")
+    assert hasattr(y, "flags")
+
+    assert np.all(
+        y.flags["profile_envelop"] == [1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 4, 1, 9]
+    )
