### 2020-01-01 838cca5
Updating valid geolocation test
It now uses the standard QCCheck class.
