+++ cotede/qctests/__init__.py
@@ -24,3 +24,4 @@ from .density_inversion import DensityInversion, densitystep
 from .stuck_value import StuckValue
 from .gradient_depthconditional import GradientDepthConditional
 from .spike_depthconditional import SpikeDepthConditional
+from .valid_geolocation import ValidGeolocation
