### 2020-01-02 7bfb145
Forgot to include gradient_depthconditional.py
