### 2020-01-06 543bdf2
Updating location at sea procedure
Using default QCCheck class and ready for tracks like TSG.
