+++ cotede/qctests/location_at_sea.py
@@ -4,12 +4,15 @@ import logging
 
 import numpy as np
 from numpy import ma
-import oceansdb
 
 from .qctests import QCCheck
 
 module_logger = logging.getLogger(__name__)
 
+try:
+    import oceansdb
+except ImportError:
+    module_logger.debug("OceansDB package is not available")
 
 def location_at_sea(data, cfg=None):
     """ Evaluate if location is at sea.
@@ -32,37 +35,35 @@ def location_at_sea(data, cfg=None):
     except:
         flag_bad = 3
 
-    assert hasattr(data, 'attributes'), "Missing attributes"
+    assert hasattr(data, 'attrs'), "Missing attributes"
 
     # Temporary solution while migrating to OceanSites variables syntax
-    if ('LATITUDE' not in data.attributes) and \
-            ('latitude' in data.attributes):
+    if ('LATITUDE' not in data.attrs) and ('latitude' in data.attrs):
                 module_logger.debug("Deprecated. In the future it will not accept latitude anymore. It'll must be LATITUDE")
-                data.attributes['LATITUDE'] = data.attributes['latitude']
-    if ('LONGITUDE' not in data.attributes) and \
-            ('longitude' in data.attributes):
+                data.attrs['LATITUDE'] = data.attrs['latitude']
+    if ('LONGITUDE' not in data.attrs) and ('longitude' in data.attrs):
                 module_logger.debug("Deprecated. In the future it will not accept longitude anymore. It'll must be LONGITUDE")
-                data.attributes['LONGITUDE'] = data.attributes['longitude']
+                data.attrs['LONGITUDE'] = data.attrs['longitude']
 
-    if ('LATITUDE' not in data.attributes) or \
-            (data.attributes['LATITUDE'] == None) or \
-            ('LONGITUDE' not in data.attributes) or \
-            (data.attributes['LONGITUDE'] == None):
+    if ('LATITUDE' not in data.attrs) or \
+            (data.attrs['LATITUDE'] == None) or \
+            ('LONGITUDE' not in data.attrs) or \
+            (data.attrs['LONGITUDE'] == None):
                 module_logger.debug("Missing geolocation (lat/lon)")
                 return 0
 
-    if (data.attributes['LATITUDE'] > 90) or \
-            (data.attributes['LATITUDE'] < -90) or \
-            (data.attributes['LONGITUDE'] > 360) or \
-            (data.attributes['LONGITUDE'] < -180):
+    if (data.attrs['LATITUDE'] > 90) or \
+            (data.attrs['LATITUDE'] < -90) or \
+            (data.attrs['LONGITUDE'] > 360) or \
+            (data.attrs['LONGITUDE'] < -180):
                 return flag_bad
 
     try:
         ETOPO = oceansdb.ETOPO()
         etopo = ETOPO['topography'].extract(
                 var='height',
-                lat=data.attributes['LATITUDE'],
-                lon=data.attributes['LONGITUDE'])
+                lat=data.attrs['LATITUDE'],
+                lon=data.attrs['LONGITUDE'])
         h = etopo['height']
 
         flag = np.zeros(h.shape, dtype='i1')
@@ -75,44 +76,40 @@ def location_at_sea(data, cfg=None):
         return 0
 
 
-def get_bathymetry(lat, lon):
+def get_bathymetry(lat, lon, resolution='5min'):
     """Interpolate bathymetry from ETOPO
 
        For a given (lat, lon), interpolates the bathymetry from ETOPO
     """
     assert np.shape(lat) == np.shape(lon), "Lat & Lon shape mismatch"
-    assert np.shape(lat) == ()
 
-    ETOPO = oceansdb.ETOPO()
+    db = oceansdb.ETOPO(resolution=resolution)
 
[…]
