### 2020-01-07 #1985 @danielskatz
no, sorry, my mistake - we first need an editor - @Kevin-Mattheus-Moerman started handling this as the rotating associate editor-in-chief, not as the editor
