### 2020-01-07 #1985 @kthyng
@castelao I see that your package is self-described as "old" and doesn't use more modern packages like pandas and xarray. As a potential user of your package but also a pythonista who is trying to stay modern in my code, can you justify this more to me and others? If there is a small amount of work that could make your package at least able to tie into these other packages, that would be hugely useful to the community using your package. A number of other packages have been updating over time to be able to, for example, recognize and use data when it is input as a pandas dataframe or xarray dataarray/dataset.
