### 2020-01-08 #1985 @castelao
@kthyng, thanks for raising this point. I see CoTeDe as an old package but not outdated. Yes, I also use and appreciate pandas and xarray, so I intend to provide a direct connection to them, but that will be an optional resource, so I do not force more dependencies than required. A few groups and packages already use CoTeDe, and I do not want to impose more dependencies than necessary. For instance, to Quality Control WOD on its binary files, QC a PostgreSQL database or QC directly cnv (CTD) files (https://github.com/castelao/seabird/blob/master/docs/notebooks/QualityControl.ipynb), there is no need for more.

Note that at some point, I changed my standard from .attributes to .attrs to match xarray so that CoTeDe could read directly from a dataset (xarray object). I still need to make a few adjustments to be fully transparent, but I think it is better to keep the CoTeDe data model simple so that a dictionary of np.array or an xr.dataset would both work.
