### 2020-01-08 #1985 @koldunovn
@kthyng Thank you for the nomination, but I am afraid I can't do it fast due to the huge amount of work and family commitments I have at the moment. Moreover, observations are not really my thing anymore, I am in the modeling business for over 15 years now, and the set of problems is quite different. I suggest to contact @ocefpaf @ledm @zklaus . They have experience in both python and oceanography.
