### 2020-01-08 #1985 @kthyng
@castelao Thank you for your response. Another possibility would be to check to see if pandas or xarray are installed and then extend possible behavior, or something similar. This happens in cartopy for example, which uses different algorithms if shapely is installed. I agree that simple is good, but it is also good to keep modernizing instead of being held back by old code that users have. Something to keep in mind, anyway.
