### 2020-01-08 #1985 @kthyng
Thanks for letting me know @koldunovn!

Hi @ocefpaf @ledm @zklaus! You've all been suggested as possible reviewers for this submission to the Journal for Open Source Software, which is largely code-review based for software with scientific applications. Are you interested in reviewing this submission? We ask for reviews within about 3 weeks if possible, and you can find more information about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

In total, we need two reviewers. Thanks.
