### 2020-01-09 130ce5d
Updating woa normbias comparison
Using default QCCheckVar class and ready for tracks like TSG.
