### 2020-01-10 06c3257
A new test for inheritance priority
Verify if the inheritance priority respects the order as they appear.
