### 2020-01-10 68e909b
Defining a new structure for QC config descriptor
