### 2020-01-14 01bca37
Removing Python-2 from Travis
