### 2020-01-14 1f6e8b6
Generic classes for tests
QCCheckVar is a variation of QCCheck that requires a varname. For
instance, a gradient check is for a specific variable like temperature.
