### 2020-01-14 c6d6011
Tests update due to new config structure
