### 2020-01-14 92027ba
Removing redundancy
