+++ cotede/qc.py
@@ -193,15 +193,6 @@ class ProfileQC(object):
             except:
                 module_logger.warning("Fail on valid_speed")
 
-        if 'global_range' in cfg:
-            y = GlobalRange(self.input, v, cfg['global_range'], autoflag=True)
-
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
-
         if 'regional_range' in cfg:
             module_logger.warning(
                     "Sorry, I'm no ready to evaluate regional_range()")
@@ -209,76 +200,6 @@ class ProfileQC(object):
         if 'pressure_increasing' in cfg:
             module_logger.warning(
                     "Sorry, I'm no ready to evaluate pressure_increasing()")
-
-        if 'profile_envelop' in cfg:
-            y = ProfileEnvelop(
-                    self.input, v, cfg['profile_envelop'], autoflag=True)
-
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
-
-        if 'constant_cluster_size' in cfg:
-            y = ConstantClusterSize(
-                    self.input, v, cfg['constant_cluster_size'], autoflag=True)
-
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
-
-        if 'gradient' in cfg:
-            y = Gradient(self.input, v, cfg['gradient'], autoflag=True)
-
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
-
-        if 'gradient_depthconditional' in cfg:
-            y = GradientDepthConditional(self.input,
-                                         v,
-                                         cfg['gradient_depthconditional'],
-                                         autoflag=True)
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
-
-        if 'spike' in cfg:
-            y = Spike(self.input, v, cfg['spike'], autoflag=True)
-
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
-
-
-        if 'spike_depthconditional' in cfg:
-            y = SpikeDepthConditional(self.input,
-                                         v,
-                                         cfg['spike_depthconditional'],
-                                         autoflag=True)
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
-
-        if 'stuck_value' in cfg:
-            y = StuckValue(self.input, v, cfg['stuck_value'], autoflag=True)
-
-            if self.saveauxiliary:
-                for f in y.features.keys():
-                    self.features[v][f] = y.features[f]
-            for f in y.flags:
-                self.flags[v][f] = y.flags[f]
 
         if 'grey_list' in cfg:
             module_logger.warning("Sorry I'm not ready to evaluate grey_list()")
@@ -295,8 +216,27 @@ class ProfileQC(object):
             module_logger.warning(
                     "Sorry I'm not ready to evaluate deepest_pressure()")
 
[…]
