### 2020-01-14 0ca2a1c
Only evaluates common tests if defined in config
