### 2020-01-14 #1985 @zklaus
@kthyng, thanks for the invite, but while I have some python experience, I don't feel qualified on the ocean data quality aspect, so I have to decline.
