### 2020-01-15 #1985 @seagrinch
@kthyng Like many of us, I'm a bit overloaded in prep for Ocean Sciences, but I can try to take a look. I am intrigued by this repo and curious to see how easy it is to utilize.
