### 2020-01-15 95cb935
BUGFIX #37, outdated notebook on fuzzy logic.
