+++ AUTHORS.rst
@@ -11,3 +11,4 @@ Contributors
 ------------
 
 * Luiz Irber <luiz.irber@gmail.com>  
+* @lidong74; Reported bug #37, outdated notebook on fuzzy logic.
