### 2020-01-16 #1985 @gmaze
Hi @karthik
Yes I'd like to review this paper,
But like many I'll be very busy to the end of February because of the Ocean Sciences meeting (and many other commitments)
So I can't guarantee to finish the review within 3 weeks.
g
