### 2020-01-17 797e594
Deprecate cotede_dir in favor to cotederc()
