### 2020-01-17 f53a5f4
cotederc for windows path standard
Nobody complained but I believe that explicit paths should be a problem
for windows that uses a different standard. A generic path can be
constructed with os.path
