### 2020-01-23 8582eba
Regional Range
Equivalent to global range test but with a defined region where it is
applicable.
