### 2020-01-23 05f6179
Missing thresholds for depth conditional spike
