+++ docs/source/qctests.rst
@@ -253,6 +253,8 @@ EuroGOOS
 |  - > 500           |            |        | - 3.0 C     | - 0.5    |
 +--------------------+------------+--------+-------------+----------+
 | Spike Cond.        |  1         | 4      |             |          |
+|  - < 500           |            |        | - 6.0 C     | - 0.9    |
+|  - > 500           |            |        | - 2.0 C     | - 0.3    |
 +--------------------+------------+--------+-------------+----------+
 | `Climatology`_     |  1         |        |                        |
 +--------------------+------------+--------+-------------+----------+
