### 2020-01-23 #1985 @kthyng
Hi @gmaze, that is certainly a trend right now!

@callumrollo I see you are on the JOSS reviewer list with an expertise in Marine Technology. Do you work with oceanographic data? This submission to JOSS is about quality control of ocean data so I'm looking for reviewers with this background plus Python expertise. I'm hoping to move through reviews in the next 3 weeks if possible. Any interest?
