### 2020-01-24 cbc46ae
Allow tests to have own default flags.
For instance, 3 is a better default bad_flag for climatology test.
