### 2020-01-25 #1985 @ocefpaf
@kthyng I spoke with Jessica Austin (@jessicaaustin), from Axiom DataScience, and she is willing to review it. Jessica a lot of experience with scientific Software development applied to oceanography and is the lead developer for the IOOS QA/QC QARTOD implementation.
