+++ cotede/qctests/__init__.py
@@ -26,3 +26,4 @@ from .stuck_value import StuckValue
 from .gradient_depthconditional import GradientDepthConditional
 from .spike_depthconditional import SpikeDepthConditional
 from .valid_geolocation import ValidGeolocation
+from .deepest_pressure import DeepestPressure
