+++ docs/source/qctests.rst
@@ -47,157 +47,6 @@ Some procedures also provide a continuous scale usually representing the probabl
 For details on that, please check the description of the specific check.
 
 
-=====
-Tests
-=====
-
-These are the tests available in CoTeDe.
-Most of the QC recommended guides follow simmilar procedures with small variations, as described below when relevant.
-
-
-Valid Date
-~~~~~~~~~~
-
-Check if there is a valid date and time associated with the measurement.
-
-.. _Argo_valid_date:
-
-For Argo, the year also must be later than 1997.
-
-Valid Position
-~~~~~~~~~~~~~~
-
-.. _Argo_valid_position:
-
-Check if there is a valid position associated with the measurement. It should have a latitude between -90 and 90, and a longitude between -180 and 360.
-
-Location at Sea
-~~~~~~~~~~~~~~~
-
-.. _Argo_on_land:
-
-Check if the position is at sea, which is evaluated using ETOPO1, a bathymetry with resolution of 1 minute.
-The point is considered at sea if the interpolated position has a negative vertical level.
-
-This test implicitly requires the data to pass the `Valid Position`_ test.
-
-.. _test-deepest-pressure:
-
-Deepest Pressure
-~~~~~~~~~~~~~~~~
-
-Check for each measurement if the reference pressure (depth) is deeper than the operational limit for that sensor/platform. For instance, the Argo Solo-II operates up to 2000m while the Deep Solo goes up to 6000m. Measurements deeper than that suggest a bad vertical position.
-
-Reference: Argo QC manual 2.9.1
-
-Global Range
-~~~~~~~~~~~~
-
-This test evaluates if the measurement is a possible value in the ocean in normal conditions.
-The thresholds used are extreme values, wide enough to accommodate all possible values and do not discard uncommon, but possible, conditions.
-
-Regional Range
-~~~~~~~~~~~~~~
-
-The regional Range test is equivalent to the Global Range but has a domain where it is applicable. Good examples are the Mediterranean Sea and the Red Sea, where the feasible range is more restrictive than required for the Global Range.
-
-This test requires the Python package Shapely to read a polygon geometry and evaluate which positions are within that domain.
-
-Reference: Argo QC manual
-
-Digit Rollover
-~~~~~~~~~~~~~~~
-
-Every sensor has a limit of bits available to store the sample value, with this limit planned to cover the possible range.
-A spurious value over the bit range would be recorded as the scale rollover, resulting in a misleading value inside the possible scale.
-This test identifies extreme jumps on consecutive measurements, that are wider than expected, suggesting a rollover error.
-
-The difference on consecutive measurements must be smaller or equal to the threshold to be approved.
-
-Gradient
-~~~~~~~~
-
-  This test compares
-
-    .. math::
-
-       X_i = \left| V_i - \frac{V_{i+1} + V_{i-1}}{2} \right|
-
-Spike
-~~~~~
-
-.. math::
-
-   X_i = \left| V_i - \frac{V_{i+1} + V_{i-1}}{2} \right| - \left| \frac{V_{i+1} - V_{i-1}}{2} \right|
-
-Tukey 53H
-~~~~~~~~~
-
-This method to detect spikes is based on the procedure initially proposed by Goring & Nikora 2002 for Acoustic Doppler Velocimeters, and similar to the one adopted by Morello 2011.
-It takes advantage of the robustness of the median to create a smoother data series, which is then compared with the observation.
-This difference is normalized by the standard deviation of the observed data series after removing the large--scale variability.
-
[…]
