+++ docs/source/qctest/tests.rst
@@ -1,0 +1,209 @@
+*************************
+Tests for Quality Control
+*************************
+
+These are the tests available in CoTeDe.
+Most of the QC recommended guides follow simmilar procedures with small variations, as described below when relevant.
+
+.. _test-valid-date:
+
+Valid Date
+~~~~~~~~~~
+
+Check if there is a valid date and time associated with the measurement.
+
+.. _Argo_valid_date:
+
+For Argo, the year also must be later than 1997.
+
+For underwater gliders, the year must be later than 1998.
+
+.. _test-valid-position:
+
+Valid Position
+~~~~~~~~~~~~~~
+
+Check if there is a valid position associated with the measurement.
+It should have a latitude between -90 and 90, and a longitude between -180 and 360.
+
+.. _GTSPP-valid-position:
+
+GTSPP restricts the longitude range to between -180 and 180 degrees.
+
+.. _Argo_valid_position:
+
+.. _test-location-at-sea:
+
+Location at Sea
+~~~~~~~~~~~~~~~
+
+Check if the position is at sea by using a bathymetry database.
+If not specified, it is used ETOPO1, a bathymetry with resolution of 1 minute.
+This test implies approval on :ref:`test-valid-position`.
+
+.. _GTSPP-at-sea:
+
+If the position is determined at sea, GTSPP also evaluates the sounding, if present, but that is redundant to the Sounding Test, so this part is neglected in CoTeDe.
+
+.. _Argo_on_land:
+
+.. _test-global-range:
+
+Global Range
+~~~~~~~~~~~~
+
+This test evaluates if the measurement is a possible value in the ocean in normal conditions.
+The thresholds used are extreme values, wide enough to accommodate all possible values and do not discard uncommon, but possible, conditions.
+
+.. _test-regional-range:
+
+Regional Range
+~~~~~~~~~~~~~~
+
+This test is equivalent to the :ref:`test-global-range` but with a horizontal domain where it is applicable. Good examples are the Mediterranean Sea and the Red Sea, where the feasible range is more restrictive than required for the Global Range.
+
+Regional Range was first introduced in the GTSPP manual 1990.
+
+.. note::
+
+    This test requires the Python package
+    `Shapely <https://github.com/Toblerity/Shapely>`_ to read a polygon
+    geometry and evaluate which positions are within that domain.
+
+.. _GTSPP-regional-range:
+
+GTSPP also requires the maximum depth to be less than 5200m in the Mediterranean Sea, and less than 3500m in the Red Sea.
+
+.. _test-profile-envelop:
+
+Profile Envelop
+~~~~~~~~~~~~~~~
+
+This test is equivalent to the :ref:`test-global-range` but with a vertical domain where it is applicable, i.e. it defines the acceptable range per layer.
+For instance, GTSPP defines that deeper than 1100m up to 3000m the acceptable temperature range is from -1.5C to 18C.
+
+Profile Envelop was first introduced in the GTSPP manual 1990.
+
+.. _test-gradient:
+
+Gradient
+~~~~~~~~
+
+  This test compares
+
+    .. math::
+
+       X_i = \left| V_i - \frac{V_{i+1} + V_{i-1}}{2} \right|
+
+.. _test-gradient-cond:
+
+Gradient Depth Conditional
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. _test-spike:
+
+Spike
+~~~~~
+
+.. math::
+
+   X_i = \left| V_i - \frac{V_{i+1} + V_{i-1}}{2} \right| - \left| \frac{V_{i+1} - V_{i-1}}{2} \right|
+
+.. _test-spike-cond:
+
+Spike Depth Conditional
+~~~~~~~~~~~~~~~~~~~~~~~
+
+.. _test-tukey53H:
+
[…]
