### 2020-01-30 #1985 @kthyng
@evanleeturner JOSS is the Journal for Open Source Software, which involves largely code-review based for software with research applications. Are you interested in reviewing this submission? We ask for reviews within about 3 weeks if possible, and you can find more information about the process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
I'm looking for one more reviewer! Thanks for your consideration.
