### 2020-02-03 #2063 @whedon
[REVIEW]: A Framework to Quality Control Oceanographic Data
**Submitting author:** @castelao (<a href="http://orcid.org/0000-0002-6765-0708">Guilherme Castelao</a>)
**Repository:** <a href="https://github.com/castelao/CoTeDe" target ="_blank">https://github.com/castelao/CoTeDe</a>
**Version:** v0.21.3
