### 2020-02-03 #1985 @kthyng
@jessicaaustin and @evanleeturner Please to go [#2063](https://github.com/openjournals/joss-reviews/issues/2063) for the actual review. You'll find a reviewer checklist for each of you and some instructions. Please open new issues in the software repo and link back to the review issue for items that arise, and let me know if you have any questions!
