### 2020-02-07 40e569b
Typo on DOI of Argo QC manual
