+++ joss/paper.bib
@@ -14,7 +14,7 @@
   year = "2015",
   month = "December",
   note = "Version 3.0",
-  doi = " http://dx.doi.org/10.13155/33951",
+  doi = "10.13155/33951",
 }
 
 @manual{EGOOS:2010,
