### 2020-02-08 4941772
Monotonic z test (increasing depth/pressure)
A generic vertical displacement test that can switch between pressure,
depth, or other user choice so it attends multiple variations like
the increasing pressure from Argo or the increasing depth from GTSPP.
