+++ docs/source/qctests.rst
@@ -139,7 +139,7 @@ Argo (Incomplete)
 +-------------------------------------------------------+------------+--------+-------------+----------+
 | :ref:`test-regional-range`                            |            |        |                        |
 +-------------------------------------------------------+------------+--------+-------------+----------+
-| Pressure increasing test                              |            |        |                        |
+| :ref:`Pressure increasing <Argo_increasing-pressure>` |            |        |                        |
 +-------------------------------------------------------+------------+--------+-------------+----------+
 | :ref:`test-spike`                                     |            |        |                        |
 +-------------------------------------------------------+------------+--------+-------------+----------+
