### 2020-02-09 #1985 @castelao
@jessicaaustin and @evanleeturner, thank you for accepting to review my submission.

@kthyng, I do appreciate your effort and time to keep this review moving forward considering this pre-OSM busy times. Thanks!

@gmaze and @seagrinch, I would be very much interested in hearing any comments or questions that you might have. Please feel free to contact me by opening an issue in the [CoTeDe repository](https://github.com/castelao/CoTeDe) or directly by email. Thanks!
