### 2020-02-28 40b11bd
contributing updates
https://github.com/openjournals/joss-reviews/issues/2063
