+++ CONTRIBUTING.rst
@@ -67,6 +67,7 @@ Ready to contribute? Here's how to set up `cotede` for local development.
     $ mkvirtualenv cotede
     $ cd cotede/
     $ python setup.py develop
+    $ pip install -r test-requirements.txt
 
 4. Create a branch for local development::
 
@@ -76,11 +77,9 @@ Ready to contribute? Here's how to set up `cotede` for local development.
 
 5. When you're done making changes, check that your changes pass flake8 and the tests, including testing other Python versions with tox::
 
-    $ flake8 cotede tests
+    $ py.test cotede --flake8 -m flake8
     $ py.test tests
     $ tox
-
-   To get flake8 and tox, just pip install them into your virtualenv.
 
 6. Commit your changes and push your branch to GitHub::
 
