### 2020-02-28 castelao/CoTeDe#39 @jessicaaustin
Add conda-forge packages for CoTeDe and oceansdb
Our team uses conda and conda-forge to manage dependencies for our python projects. Looks like neither CoTeDe nor oceansdb are in conda-forge. Looks like [GSW is in there](https://github.com/conda-forge/gsw-feedstock).
