### 2020-02-28 castelao/CoTeDe#40 @jessicaaustin
Resolve flake8 failures and/or add flake8-ignore
The CONTRIBUTING docs recommend running flake8. However when I tried this, there were a lot of flake8 failures. I recommend resolving those and/or adding a list of ignored tests to flake8-ignore. Otherwise, as a contributor I'd just ignore the output of these tests.
