### 2020-02-28 castelao/CoTeDe#40 @jessicaaustin
Output when I ran this:

```
(cotede37) jaustin@merlin:~/dev/castelao/CoTeDe$ pytest cotede --flake8 -m flake8
======================================================================================== test session starts =========================================================================================
rootdir: /home/jaustin/dev/castelao/CoTeDe
plugins: xdist-1.31.0, flake8-1.0.4, forked-1.1.3
collected 40 items

cotede/__init__.py F                                                                                                                                                                           [  2%]
cotede/misc.py F                                                                                                                                                                               [  5%]
cotede/qc.py F                                                                                                                                                                                 [  7%]
cotede/anomaly_detection/__init__.py F                                                                                                                                                         [ 10%]
cotede/anomaly_detection/anomaly_detection.py F                                                                                                                                                [ 12%]
cotede/fuzzy/__init__.py F                                                                                                                                                                     [ 15%]
cotede/fuzzy/defuzz.py F                                                                                                                                                                       [ 17%]
cotede/fuzzy/fuzzy_core.py F                                                                                                                                                                   [ 20%]
cotede/fuzzy/membership_functions.py F                                                                                                                                                         [ 22%]
cotede/humanqc/__init__.py F                                                                                                                                                                   [ 25%]
cotede/humanqc/humaneval.py F                                                                                                                                                                  [ 27%]
cotede/qctests/__init__.py F                                                                                                                                                                   [ 30%]
cotede/qctests/anomaly_detection.py F                                                                                                                                                          [ 32%]
cotede/qctests/bin_spike.py F                                                                                                                                                                  [ 35%]
cotede/qctests/cars_normbias.py F                                                                                                                                                              [ 37%]
cotede/qctests/constant_cluster_size.py s                                                                                                                                                      [ 40%]
cotede/qctests/cum_rate_of_change.py F                                                                                                                                                         [ 42%]
cotede/qctests/density_inversion.py F                                                                                                                                                          [ 45%]
cotede/qctests/descentPrate.py F                                                                                                                                                               [ 47%]
cotede/qctests/digit_roll_over.py F                                                                                                                                                            [ 50%]
cotede/qctests/fuzzylogic.py F                                                                                                                                                                 [ 52%]
cotede/qctests/global_range.py F                                                                                                                                                               [ 55%]
cotede/qctests/gradient.py F                                                                                                                                                                   [ 57%]
cotede/qctests/gradient_depthconditional.py s                                                                                                                                                  [ 60%]
cotede/qctests/location_at_sea.py F                                                                                                                                                            [ 62%]
cotede/qctests/morello2014.py F                                                                                                                                                                [ 65%]
cotede/qctests/possible_speed.py F                                                                                                                                                             [ 67%]
cotede/qctests/profile_envelop.py F                                                                                                                                                            [ 70%]
cotede/qctests/qctests.py F                                                                                                                                                                    [ 72%]
cotede/qctests/rate_of_change.py F                                                                                                                                                             [ 75%]
cotede/qctests/spike.py F                                                                                                                                                                      [ 77%]
cotede/qctests/spike_depthconditional.py F                                                                                                                                                     [ 80%]
cotede/qctests/stuck_value.py s                                                                                                                                                                [ 82%]
cotede/qctests/tukey53H.py F                                                                                                                                                                   [ 85%]
cotede/qctests/valid_geolocation.py F                                                                                                                                                          [ 87%]
cotede/qctests/woa_normbias.py F                                                                                                                                                               [ 90%]
cotede/utils/__init__.py F                                                                                                                                                                     [ 92%]
cotede/utils/config.py F                                                                                                                                                                       [ 95%]
cotede/utils/profilescollection.py F                                                                                                                                                           [ 97%]
cotede/utils/utils.py F                                                                                                                                                                        [100%]

============================================================================================== FAILURES ==============================================================================================
____________________________________________________________________________________________ FLAKE8-check ____________________________________________________________________________________________
/home/jaustin/dev/castelao/CoTeDe/cotede/__init__.py:7:1: F401 'cotede.qc' imported but unused

____________________________________________________________________________________________ FLAKE8-check ____________________________________________________________________________________________
/home/jaustin/dev/castelao/CoTeDe/cotede/misc.py:38:9: E265 block comment should start with '# '
/home/jaustin/dev/castelao/CoTeDe/cotede/misc.py:38:80: E501 line too long (80 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/misc.py:44:80: E501 line too long (84 > 79 characters)

____________________________________________________________________________________________ FLAKE8-check ____________________________________________________________________________________________
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:4:1: F401 'pkg_resources' imported but unused
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:6:1: F401 'os.path.basename' imported but unused
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:8:1: F401 'json' imported but unused
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:14:1: F403 'from cotede.qctests import *' used; unable to detect undefined names
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:26:13: E128 continuation line under-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:29:28: W605 invalid escape sequence '\d'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:30:31: W605 invalid escape sequence '\d'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:31:28: W605 invalid escape sequence '\d'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:46:9: E722 do not use bare 'except'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:65:13: E265 block comment should start with '# '
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:82:45: E127 continuation line over-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:126:29: E127 continuation line over-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:126:80: E501 line too long (81 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:128:29: E127 continuation line over-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:135:17: F405 'LocationAtSea' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:170:80: E501 line too long (80 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:173:17: F405 'ValidGeolocation' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:190:29: F405 'possible_speed' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:191:13: E722 do not use bare 'except'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:203:80: E501 line too long (80 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:218:30: F405 'Bin_Spike' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:219:34: F405 'CARS_NormBias' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:220:42: F405 'ConstantClusterSize' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:221:39: F405 'CumRateOfChange' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:222:36: F405 'DigitRollOver' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:223:33: F405 'GlobalRange' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:224:29: F405 'Gradient' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:225:46: F405 'GradientDepthConditional' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:226:36: F405 'ProfileEnvelop' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:227:35: F405 'RateOfChange' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:228:26: F405 'Spike' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:229:43: F405 'SpikeDepthConditional' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:230:32: F405 'StuckValue' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:231:34: F405 'Tukey53H' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:232:33: F405 'WOA_NormBias' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:244:9: E265 block comment should start with '# '
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:259:21: F405 'DensityInversion' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:268:13: E722 do not use bare 'except'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:271:9: E265 block comment should start with '# '
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:281:34: E271 multiple spaces after keyword
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:286:17: E722 do not use bare 'except'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:288:45: F405 'spike' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:290:48: F405 'gradient' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:293:33: F405 'constant_cluster_size' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:295:53: F405 'tukey53H_norm' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:295:80: E501 line too long (80 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:297:54: F405 'rate_of_change' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:297:80: E501 line too long (82 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:299:29: F405 'WOA_NormBias' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:301:33: E127 continuation line over-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:303:29: F405 'CARS_NormBias' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:305:33: E127 continuation line over-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:308:80: E501 line too long (89 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:311:21: E127 continuation line over-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:311:21: F405 'anomaly_detection' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:317:44: F405 'morello2014' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:321:27: E271 multiple spaces after keyword
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:326:17: E722 do not use bare 'except'
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:327:80: E501 line too long (81 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:329:43: F405 'fuzzylogic' may be undefined, or defined from star imports: cotede.qctests
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:368:21: E128 continuation line under-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/qc.py:368:53: E225 missing whitespace around operator

____________________________________________________________________________________________ FLAKE8-check ____________________________________________________________________________________________
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/__init__.py:4:1: F401 '.anomaly_detection.estimate_anomaly' imported but unused
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/__init__.py:5:1: F403 'from .anomaly_detection import *' used; unable to detect undefined names
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/__init__.py:5:1: F401 '.anomaly_detection.*' imported but unused

____________________________________________________________________________________________ FLAKE8-check ____________________________________________________________________________________________
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:19:80: E501 line too long (101 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:21:80: E501 line too long (489 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:23:80: E501 line too long (230 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:25:80: E501 line too long (91 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:39:1: F401 'cotede.misc.combined_flag' imported but unused
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:70:21: E128 continuation line under-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:75:80: E501 line too long (86 > 79 characters)
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:114:17: E128 continuation line under-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:159:26: E203 whitespace before ','
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:159:33: E702 multiple statements on one line (semicolon)
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:161:55: E261 at least two spaces before inline comment
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:161:55: E262 inline comment should start with '# '
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:205:9: F821 undefined name 'pylab'
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:206:9: F821 undefined name 'pylab'
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:209:13: E128 continuation line under-indented for visual indent
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:223:5: E265 block comment should start with '# '
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:227:5: E265 block comment should start with '# '
/home/jaustin/dev/castelao/CoTeDe/cotede/anomaly_detection/anomaly_detection.py:229:5: E265 block comment should start with '# '
[…]
