### 2020-02-28 castelao/CoTeDe#41 @jessicaaustin
Add API documentation
I might be missing something, but I don't see docs for the API at https://cotede.readthedocs.io/en/latest/readme.html

Looks like the code as docstrings for most classes and test methods, so this is probably just a matter of configuring sphinx to produce the docs.
