### 2020-03-03 7b855ea
BUGFIX #40, do not recommend flake8 anymore.
I've been using black to format the code. I'll not refuse contributions
just due to bad formating.
