+++ CONTRIBUTING.rst
@@ -74,13 +74,12 @@ Ready to contribute? Here's how to set up `cotede` for local development.
 
    Now you can make your changes locally.
 
-5. When you're done making changes, check that your changes pass flake8 and the tests, including testing other Python versions with tox::
+5. When you're done making changes, check that your changes pass the tests, including testing other Python versions with tox::
 
-    $ flake8 cotede tests
     $ py.test tests
     $ tox
 
-   To get flake8 and tox, just pip install them into your virtualenv.
+   To get tox, just pip install them into your virtualenv.
 
 6. Commit your changes and push your branch to GitHub::
 
