### 2020-03-03 #2063 @evanleeturner
### Conflict of interest

I do not have a conflict of interest in this review.  I do maintain internal (non-public) code repository @wdft-coastal that does contain some of the similar functions as presented in the current code but are specific for my dataset and organization.

### Overall thoughts:
I am extremely pleased with the premise of this package for organizing QC procedures and being able to run multiple tests.  This is a high priority item for my organization and my current data sets, as our own QAQC methods often can conflict with other agencies that follow different QAQC procedures. In the last few months I have been preparing code that can mimic QARTOD style tests on our own data internally, but I wanted the ability to apply alternative test schemes and to skip tests that are not appropriate for our data.  Having a maintained package that accomplishes this task is not only helpful but it also greatly increases your data confidence as now the QC methods are maintained by other groups in a transparent mechanism.  I fully expect to be implementing this package into our own website and data products at TWDB.

I share reviewer one’s sentiment about the philosophy of leveraging other data libraries such as netCDF or pandas.  Although our own internal datasets are SQL we do make use of pandas where possible for data manipulation and in the future we will likely be forced to move to netCDF to process very large datasets for QAQC.  This functionality is not a show-stopper for the package, but it is a severe limitation for its acceptance in the community.

### Review
Items marked **BLOCKER** below should be resolved before acceptance. Everything else is a recommendation.
Paper:
•	The paper itself was succinct, well-written, and very appropriate to the code.
•	**Recommended:**  Please consider adding the functionality of swapping methods for calculating seawater: https://github.com/castelao/CoTeDe/issues/43
•	**Recommended:** mirror reviewer one that I don’t know how to pronounce CoTeDe! 😊
Documentation:
•	**BLOCKER**: Please also complete the issue that was brought by reviewer one concerning the API docs as this would also be extremely helpful for me to implement this code: https://cotede.readthedocs.io/ . castelao/CoTeDe#41
Installation
•	I was able to get the package running.
Contributing
•	I was able to run the tests.
Performance
•	There were no performance tests or statements regarding performance.
•	I did not have an issue with performance with my own tests, but I also have very small datasets at the moment.  Additionally, if I were to run this package in a production environment, the performance hit would be negated because our data would be checked as it inserted into our database which would be a heftier cost of CPU cycles.  If I were going to run these tests on the entire database, I would do so offline or during non-peak hours on the production systems in any case.
