### 2020-03-04 castelao/CoTeDe#43 @castelao
Nice! Thanks. I agree that it can cause some differnce, but I have to think about it. It might not be the job of CoTeDe to decide which procedure to use to obtain the salinity, but restrict on evaluating the data that is given. I believe that I currently use GSW only to estimate the density when it's not available, to identify density inversions.

CoTeDe was originally a full CTD processing tool, from parsing different formats, QC, and exporting netCDFs, but as it grew too large, I split that in multiple packages. For instance, now [pySeabird](https://github.com/castelao/seabird) deals with some formats of CTD, and calls CoTeDe to QC the data.
