### 2020-03-04 cabdf49
Download climatologies before running Travis tests.
I removed that some time ago to speed up the tests, but there is no way
to test for salinity as well without downloading the climatologies. It
downloads only once, so it is not an issue for the users.
