### 2020-03-09 7db6f42
New function utils.list_cfgs()
It returns a list of all QC procedures, the built-in plus the user local
collection.
