### 2020-03-10 #1985 @callumrollo
Hi @kthyng  sorry for missing this, I was at sea for 6 weeks with no internet. Just the kind of thing I would be happy to review in future though. @castelao I'll have a proper dig into the code once I'm caught up on post cruise admin. Looks like an awesome tool!
