### 2020-03-11 da7d7af
Improving documentation
Addressing issue #41
