+++ cotede/__init__.py
@@ -5,3 +5,4 @@ __email__ = 'guilherme@castelao.net'
 __version__ = '0.21.0'
 
 from cotede import qc
+from cotede.qc import ProfileQC, ProfileQCed
