+++ docs/source/conf.py
@@ -12,13 +12,16 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import sys
+from datetime import datetime
 import os
+import sys
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #sys.path.insert(0, os.path.abspath('.'))
+
+import cotede
 
 # -- General configuration ------------------------------------------------
 
@@ -29,9 +32,14 @@ import os
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.imgmath'
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.imgmath",
+    "sphinx.ext.napoleon",
 ]
+
+autosummary_generate = True
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
@@ -47,16 +55,16 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'CoTeDe'
-copyright = u'2015, Guilherme Castelão'
+copyright = "2006-{}, Guilherme Castelão".format(datetime.now().year)
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The short X.Y version.
-version = '0.21'
+version = ".".join(cotede.__version__.split(".")[:2])
 # The full version, including alpha/beta/rc tags.
-release = '0.21.0'
+release = cotede.__version__
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
