+++ cotede/utils/__init__.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python
 
 from .utils import *
+from .utils import cotederc
 from cotede.utils.profilescollection import ProfilesQCCollection
 from cotede.utils.profilescollection import ProfilesQCPandasCollection
-from .config import load_cfg
+from .config import load_cfg, list_cfgs
 
 import logging
