+++ docs/source/generated/cotede.ProfileQC.rst
@@ -1,0 +1,25 @@
+cotede.ProfileQC
+================
+
+.. currentmodule:: cotede
+
+.. autoclass:: ProfileQC
+
+
+   .. automethod:: __init__
+
+
+   .. rubric:: Methods
+
+   .. autosummary::
+
+      ~ProfileQC.__init__
+      ~ProfileQC.keys
+
+
+   .. rubric:: Attributes
+
+   .. autosummary::
+
+      ~ProfileQC.attrs
+      ~ProfileQC.features

[…]
