### 2020-03-13 #2063 @evanleeturner
It won't let me change the box when I click it - nothing happens.
