### 2020-03-14 #2063 @arfon
Whedon's invite functionality is definitely making the invite correctly:

@evanleeturner - are you sure you're logged in with the correct GitHub account when clicking the link?
