### 2020-03-16 #2063 @castelao
@jessicaaustin , thanks for the fast response. Yes, my bad. The latest version of the documentation is on the dev branch:

https://cotede.readthedocs.io/en/dev/api.html
