### 2020-03-17 #2063 @evanleeturner
Looks like i was able to check off the first items...

I am ready to accept this paper!
