### 2020-03-17 #2063 @arfon
> Looks like i was able to check off the first items...

@evanleeturner - thanks for the update. Just verifying that you've completed your review? If so, could you update the checklist accordingly (if permissions are fixed for you now).
