### 2020-03-17 #2063 @evanleeturner
Yes, I am finished. When I click one box it sits there and locks the others - so just keeping diligent here and checking all of them as they open up for me.
