### 2020-03-19 #2063 @kthyng
>
> > ```
> > * Highly recommended, if possible: Include a list of groups actively using this library, to give us an idea of how widely used it is. (this could also be added to the code documentation)
> > ```
>
> Thanks, that’s a good idea. However, I believe that the choice to be listed belongs to the users. I’ll create a designated space for that on CoTeDe’s website and encourage the users to add themselves to the list.
>

As listed on the [JOSS website](https://joss.readthedocs.io/en/latest/submitting.html), authors should in their paper: "Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it."
