### 2020-03-20 d2671c4
Fixing documentation badge link
It was pointing to the project at readthedocs. Better link directly to
the latest documentation. It was certainly creating some confusion
before.
