### 2020-03-20 castelao/CoTeDe#41 @castelao
@jessicaaustin , thanks for raising this issue. I updated the documentation (https://cotede.readthedocs.io) to include the main resources that a new user should know. I'll soon finish cleaning the code of this package, and then I'll be able to use the automatic mode for a more extensive documentation coverage.
