### 2020-03-22 f124c94
Cleaning resources moved to pySeabird
Recently I moved all the CTD/TSG specific resources to pySeabird to
generalize CoTeDe even more. Eventually I'll replace these classes with
something generic.
