+++ cotede/utils/profilescollection.py
@@ -1,289 +1,0 @@
-# -*- coding: utf-8 -*-
-
-""" Functions to handle a collection of profiles
-"""
-
-import time
-# import logging
-import multiprocessing as mp
-
-import numpy as np
-from numpy import ma
-
-try:
-    from seabird.utils import make_file_list
-except:
-    pass
-import cotede.qc
-
-
-def process_profiles_serial(inputfiles, cfg=None, saveauxiliary=False,
-        verbose=True):
-#                verbose=True, logger=None):
-    """ Quality control a list of CTD files
-    """
-#    logger = logger or logging.getLogger(__name__)
-    print("This function will be removed. If interested in processing CTD files, please install the package PySeabird (pip install seabird).")
-    import time; time.sleep(5)
-
-
-    profiles = []
-    for f in inputfiles:
-        try:
-            p = cotede.qc.fProfileQC(f, cfg, saveauxiliary, verbose=verbose)
-            profiles.append(p)
-        except CNVError as e:
-            print(e.msg)
-#            logger.warn(e.msg)
-
-    return profiles
-
-
-def process_profiles(inputfiles, cfg=None, saveauxiliary=True,
-        verbose=True, timeout=60):
-    # verbose=True, timeout=60, logger=None):
-    """ Quality control a list of CTD files in parallel
-    """
-    print("This function will be removed. If interested in processing CTD files, please install the package PySeabird (pip install seabird).")
-    import time; time.sleep(5)
-    # logger = logger or logging.getLogger(__name__)
-    npes = 2 * mp.cpu_count()
-    npes = min(npes, len(inputfiles))
-    # logger.debug("Running with %s npes" % npes)
-    queuesize = 3*npes
-    # logger.debug("queue size: %s" % queuesize)
-    qout = mp.Queue(queuesize)
-    teste = []
-
-    def run_qc(inputfiles, cfg, saveauxiliary, verbose):
-        def process_file(f, cfg, saveauxiliary, verbose=verbose):
-            try:
-                if verbose is True:
-                    print("Loading: %s" % f)
-#                logger.debug("fProfileQC: %s" % f)
-                p = cotede.qc.fProfileQC(f, cfg, saveauxiliary, verbose)
-#                logger=logger)
-                attrs = [pn.attributes for pn in p.data]
-#                logger.debug("Sending profile %s to queue" % f)
-                qout.put([p, attrs], block=True)
-#                logger.debug("Sent to queue")
-            except CNVError as e:
-                print(e.msg)
-#                logger.warn(e.msg)
-
-        pool = []
-        for f in inputfiles[:npes]:
-            pool.append(mp.Process(target=process_file,
-                args=(f, cfg, saveauxiliary, verbose)))
-            pool[-1].start()
-
-        for i, f in enumerate(inputfiles[npes:]):
-            n = i % npes
-            pool[n].join(timeout)
-            if pool[n].is_alive():
-                print("timeout: %s" % pool[n])
-            pool[n].terminate()
-            pool[n] = mp.Process(target=process_file,
-                args=(f, cfg, saveauxiliary, verbose))
-            pool[n].start()
-
-        for p in pool:
-            p.join(timeout)
-            if p.is_alive():
-                print("timeout: %s" % p)
-            p.terminate()
-        print("Done evaluating.")
-
-    worker = mp.Process(target=run_qc,
-            args=(inputfiles, cfg, saveauxiliary, verbose))
-    worker.start()
-
-    profiles = []
-    while worker.is_alive() or not qout.empty():
-        if qout.empty():
-            # logger.debug("Queue is empty. I'll give a break.")
-            # print("Queue is empty. I'll give a break.")
-            time.sleep(2)
-        else:
[…]
