+++ cotede/utils/__init__.py
@@ -2,8 +2,6 @@
 
 from .utils import *
 from .utils import cotederc
-from cotede.utils.profilescollection import ProfilesQCCollection
-from cotede.utils.profilescollection import ProfilesQCPandasCollection
 from .config import load_cfg, list_cfgs
 
 import logging
