### 2020-03-22 8807c0a
A new section in the documention to list users
Although CoTeDe has been used in several institutions around the world, to my knowledge, there are no citations to CoTeDe's DOIs. I guess the community is still not used with the idea of citing software DOIs, but I hope to revert it with the JOSS' publication.

I'll maintain here a list of publications, and a list of projects using CoTeDe. Thanks to @jessicaaustin for the suggestion.
