+++ docs/source/usedby.rst
@@ -1,0 +1,29 @@
+*******
+Used by
+*******
+
+Projects
+--------
+This is an incomplete list of projects that used or have been using CoTeDe. If your project is not listed here, please let me know. Thanks!
+
+  * Support for automatic QC plus a novel approach to optimize the expert QC effort for the `International Quality Controlled Ocean Database (IQuOD) <http://www.iquod.org/>`_, since 2015.
+  * CTD hydrographic operations by the National Institute for Space Research (INPE), 2011.
+  * Thermosalinograph operations of the Atlantic Oceanographic and Meteorological Laboratory - NOAA (AOML), 2006.
+
+
+Publications
+------------
+
+To cite CoTeDe, please include the reference below in your bibliography::
+
+  @article{cotede,
+    author = {Guilherme P. Castelão},
+    title = {A Framework to Quality Control Oceanographic Data},
+    journal = {Journal of Open Source Software},
+    year = {2020},
+    month=mar,
+    publisher={The Open Journal},
+    doi={TO BE DEFINED}
+  }
+
+Who would like to be the first one to cite CoTeDe?
