### 2020-03-22 f3336d6
Adding projects that used CoTeDe
As requested by @kthyng at openjournals/joss-reviews#2063
