+++ joss/paper.md
@@ -23,7 +23,10 @@ The ocean is an intrinsically challenging environment to collect data, which mak
 Automatic QC traditionally consists of a sequence of tests that identify spurious measurements based on different criteria; thus, it is common to aggregate multiple tests to cover a more extensive possibility of problems. With that in mind, the tests of the most common QC procedures were implemented in CoTeDe in a modular fashion to facilitate expanding with new tests and to permit alternative arrangements. The user can choose from one of the built-in standard procedures, for example, the Argo recommendations [@Argo:2015], or compose a custom arrangement of tests. This freedom allows a better data assessment by fine-tuning the methods -- A Spray underwater glider operating on the California coast might benefit from slightly different thresholds than used to QC the same Spray in the Mediterranean. The custom QC procedures can be saved as JSON descriptors to be re-used or shared so that experts can develop a diverse collection of alternative procedures without necessarily requiring more coding. The implementation of such a versatile system was simplified by adopting a generic data model object that standardizes the data and metadata access, resulting in a platform-agnostic system that could equally evaluate data from an XBT or a Spray underwater glider with the appropriate tests. For instance, the concept of the World Ocean Atlas climatology comparison test [@GTSPP:2010] is the same for a profile from an XBT or a CTD, being only necessary to inform the observed temperature, geolocation, and time. Finally, CoTeDe dependencies were intentionally minimized to simplify its adoption by other packages.
 
 
-Modern oceanography relies on an unprecedented abundance of measurements that keeps increasing every day [see @WOD2018]. In the past couple of decades, traditional observational systems have been complemented by autonomous platforms like Argo, Spray underwater gliders, and Saildrones, which can provide nearly continuous data streams. Thus, QC procedures for oceanographic data must evolve to keep pace with the ever-growing dataflow. In this context, CoTeDe was developed to reduce the burden on manual expert QC with automation, which enables the expert to focus their effort on the most challenging cases. In addition, the flexibility of CoTeDe provides a framework to implement novel QC approaches such as fuzzy logic and machine learning. CoTeDe is the result of several generations of quality control systems that have evolved since 2006, and it was designed to attend not only individual scientists but also real-time operations on large data centers.
+
+
+Modern oceanography relies on an unprecedented abundance of measurements that keeps increasing every day [see @WOD2018]. In the past couple of decades, traditional observational systems have been complemented by autonomous platforms like Argo, Spray underwater gliders, and Saildrones, which can provide nearly continuous data streams. Thus, QC procedures for oceanographic data must evolve to keep pace with the ever-growing dataflow. In this context, CoTeDe was developed to reduce the burden on manual expert QC with automation, which enables experts to focus their effort on the most challenging cases. CoTeDe was designed to attend not only individual scientists but also real-time operations on large data centers while providing a flexible framework to implement novel QC approaches such as fuzzy logic and machine learning. This is the result of several generations of quality control systems, and some of the most notable applications of CoTeDe have been the QC of real-time thermosalinograph operations by the Atlantic Oceanographic and Meteorological Laboratory - NOAA (AOML) in 2006; PIRATA hydrographic cruises by the National Institute for Space Research (INPE) in 2011; and a novel approach to optimize expert manual QC by the International Quality Controlled Ocean Database (IQuOD), since 2015.
+
 
 # Acknowledgements
 
