### 2020-03-24 castelao/CoTeDe#43 @castelao
Hi @evanleeturner , thanks for your suggestion. This is indeed relevant for comparing datasets, or evaluating long time series, but, as I mentioned before, I don't think it is the job for CoTeDe. My vision for CoTeDe is to be used by other packages for specific tasks. While CoTeDe focus on QC, these overarching packages would deal with problems like you raised - a uniform data processing - and deliver what is to be quality controlled to CoTeDe. If you plan to develop that for your internal project, let me know if I can give you suggestions on how to integrate it. Cheers,
