### 2020-03-24 #2063 @castelao
Thanks to everyone involved in this process, I do appreciate your time on this.

@kthyng I believe that I addressed everything. What is the next step? Shall I do a new release and obtain a new DOI? Thanks,
