### 2020-03-25 bd03f33
Added some {}'s to preserve capitalization
And a few other tweaks
