+++ joss/paper.bib
@@ -1,11 +1,11 @@
 
 
-@incollection{WOD2018,
-  title = "World Ocean Database 2018",
+@manual{WOD2018,
+  title = "{World Ocean Database} 2018",
   author = "Boyer, T. P. and O. K. Baranova and  C. Coleman and  H. E. Garcia and  A. Grodsky and  R. A. Locarnini and  A. V. Mishonov and  C. R. Paver and  J. R. Reagan and  D. Seidov and  I. V. Smolyar and  K. Weathers and  M. M. Zweng",
   year = "2019",
-  editor = ". A. V. Mishonov, Technical Editor",
-  series = "NOAA Atlas NESDIS 87",
+  note = "A. V. Mishonov, Technical Editor",
+  series = "NOAA {Atlas} {NESDIS} 87",
 }
 
 @manual{Argo:2015,
@@ -28,7 +28,7 @@
 }
 
 @manual{GTSPP:2010,
-    title = "GTSPP Real-Time Quality Control Manual",
+    title = "GTSPP {Real-Time Quality Control Manual}",
     organization = {UNESCO--IOC},
     address ="United Nations Educational, Scientific and Cultural Organization 7, Place de Fontenoy, 75352, {P}aris 07 SP",
     edition = "First Revised Edition",
