### 2020-03-25 #2063 @kthyng
@castelao Your references need some tweaks. I put some in [here](https://github.com/castelao/CoTeDe/pull/44) which you can merge if you approve. However, I see that the "note" category from "manual" entries is not coming through. Can you play around with these entries to get that information in?
