### 2020-03-26 e69ee70
Improving/fixing bibliography records
