+++ joss/paper.bib
@@ -1,67 +1,67 @@
 
 
 @manual{WOD2018,
-  title = "{World Ocean Database} 2018",
-  author = "Boyer, T. P. and O. K. Baranova and  C. Coleman and  H. E. Garcia and  A. Grodsky and  R. A. Locarnini and  A. V. Mishonov and  C. R. Paver and  J. R. Reagan and  D. Seidov and  I. V. Smolyar and  K. Weathers and  M. M. Zweng",
-  year = "2019",
-  note = "A. V. Mishonov, Technical Editor",
-  series = "NOAA {Atlas} {NESDIS} 87",
+  title = {{World Ocean Database} 2018},
+  author = {Boyer, T. P. and O. K. Baranova and  C. Coleman and  H. E. Garcia and  A. Grodsky and  R. A. Locarnini and  A. V. Mishonov and  C. R. Paver and  J. R. Reagan and  D. Seidov and  I. V. Smolyar and  K. Weathers and  M. M. Zweng},
+  year = {2019},
+  note = {A. V. Mishonov, Technical Editor},
+  series = {{NOAA} {Atlas} {NESDIS} 87},
 }
 
 @manual{Argo:2015,
-  title = "Argo Quality Control Manual For {CTD} and Trajectory Data",
-  author = "Annie Wong and Robert Keeley and Thierry Carval and {Argo Data Management Team}",
-  year = "2015",
-  month = "December",
-  note = "Version 3.0",
-  doi = "10.13155/33951",
+  title = {Argo Quality Control Manual For {CTD} and Trajectory Data},
+  author = {Annie Wong and Robert Keeley and Thierry Carval and {Argo Data Management Team}},
+  year = {2015},
+  month = {December},
+  edition = {Version 3.0},
+  doi = {10.13155/33951},
 }
 
 @manual{EGOOS:2010,
-    title = "Recommendations for \emph{in--situ} data Near Real Time Quality Control",
-    organization = "European Global Ocean Observing System",
-    edition = "EG10.19",
-    year = "2010",
-    note = "EG10.19",
-    author = "{DATA--MEQ working group}",
-        key = {DATA--MEQ working group}
+  title = {Recommendations for in--situ data Near Real--Time Quality Control},
+  organization = {European Global Ocean Observing System},
+  edition = {EG10.19},
+  year = {2010},
+  note = {EG10.19},
+  author = {{DATA--MEQ working group}},
+  key = {DATA--MEQ working group}
 }
 
 @manual{GTSPP:2010,
-    title = "GTSPP {Real-Time Quality Control Manual}",
-    organization = {UNESCO--IOC},
-    address ="United Nations Educational, Scientific and Cultural Organization 7, Place de Fontenoy, 75352, {P}aris 07 SP",
-    edition = "First Revised Edition",
-    year = "2010",
-    note = "IOC/2010/MG/22Rev.",
-    key = "{UNESCO--IOC}, 2010",
+  title = {GTSPP {Real--Time Quality Control Manual}},
+  organization = {UNESCO--IOC},
+  address = {United Nations Educational, Scientific and Cultural Organization 7, Place de Fontenoy, 75352, {P}aris 07 SP},
+  edition = {First Revised Edition},
+  year = {2010},
+  note = {IOC/2010/MG/22Rev.},
+  key = {{UNESCO--IOC}, 2010},
 }
 
 @manual{IOC26,
-        author = "{IOC/IODE}",
-        title = " Manual of quality control procedures for validation of oceanographic data",
-        year = "1993",
-        note = "IOC Manuals and guides No.26",
+  author = {{IOC/IODE}},
+  title = {Manual of quality control procedures for validation of oceanographic data},
+  year = {1993},
[…]
