### 2020-03-26 #2063 @castelao
@kthyng I'm sorry for my lack of attention on the bibliography records and thank you very much for your PR. I did a few other adjustments.
