### 2020-03-30 66d6eed
A new version with all modifications for the JOSS paper
The modifications were mostly on documentation and cleaning.
