+++ cotede/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = 'Guilherme Castelao'
 __email__ = 'guilherme@castelao.net'
-__version__ = '0.21.0'
+__version__ = '0.21.2'
 
 from cotede import qc
 from cotede.qc import ProfileQC, ProfileQCed
