### 2020-03-30 #2063 @castelao
I just release a new version of CoTeDe and the new DOI is: 10.5281/zenodo.583710

Do I need to do anything else? Thank you
