### 2020-03-30 #2063 @kthyng
@castelao I'm still seeing that the "note" section in your .bib file isn't coming through to be printed in your paper. For example, in the .bib file for Boyer et al, there is a note for the Technical Editor, but this doesn't come through to the reference in your paper. I can try to work on this at some point to see which entry type will carry through the note field, but it might be a few days before I have a chance.
