### 2020-03-30 #2063 @castelao
@kthyng, you are already too generous with your time, I can do it myself.    Thank you.

Yes, I noticed that, but my understanding is that the lack of the note field is the result of the reference type 'manual' combined with the bibliography style used by JOSS. Different journals adopt different styles, and since this is not my expertise, I always provide the fields and let each journal style to decide how to use those fields.

One alternative is to move back from manual to incollection, as it was originally, which would allow an editor field. But notice that Mishonov is a technical editor and not a regular editor, so it would not be fully correct. My opinion is that hacking it to force those changes would go against the principle of using BibTeX. I'm more inclined to provide the fields and let the style definition decides how to use those and what to show. The fields currently shown are enough to uniquely identify this reference, and it is normal to have fields that are not shown in the reference. For instance, most styles omit the fields 'keys' and 'URL'.

If you think that it is necessary to show the technical editor, I can figure this out.
