### 2020-03-30 #2063 @kthyng
@castelao Sure, the combination of items results in what ultimately shows up in the references. The thing that is most in your control is what type of entry a given reference is entered as. I think it is a good goal to have the information about the references show up, and it doesn't taking Latex hacking so much as appropriate field/entry choice. When I looked up that reference online, it was requested to show the Technical Editor in the citation, so I would try to respect that. That wasn't the only entry that had information in the bibtex entry that wasn't showing up in the references list. Can you give it another small chunk of time to do your best to faithfully represent the references?
