### 2020-03-30 #2063 @castelao
@kthyng certainly. I just did a test, and if I replace the note field by an editor field it shows up in the PDF. The problem is that the editor field is not standard for a manual type of entry. But editor field is standard for an 'incollection' entry.

Was intentional that you changed [from 'incollection' to 'manual'](https://github.com/castelao/CoTeDe/commit/bd03f3369ab430f19a6654525ba0f0700f8863ca)? I could move it back to incollection, or I can exchange 'note' by 'editor' and keep the 'manual' type of entry. Let me know your preference.
