### 2020-03-30 #2063 @kthyng
@castelao I changed from incollection to manual when I saw your other bib entries were manual and had note fields — I assumed that meant the note fields were coming through. And as I recall, the editor field wrote "Editor" but since the person is a technical editor, not just editor, I figured a less specific field would be more appropriate to use (then keeping the language "Technical Editor").
