### 2020-03-30 #2063 @castelao
One alternative to force that output is to use:

...
series = {(A. V. Mishonov, {T}echnical {E}ditor) {NOAA} {Atlas} {NESDIS}
87},
...

Would you prefer that or incollection with editor (which would be 'Ed.'
only instead of Technical Editor)?

On Mon, Mar 30, 2020 at 2:17 PM Kristen Thyng <notifications@github.com>
wrote:

> @castelao <https://github.com/castelao> I changed from incollection to
> manual when I saw your other bib entries were manual and had note fields —
> I assumed that meant the note fields were coming through. And as I recall,
> the editor field wrote "Editor" but since the person is a technical editor,
> not just editor, I figured a less specific field would be more appropriate
> to use (then keeping the language "Technical Editor").
>
