### 2020-03-31 7720b79
Moving WOD 2018 reference back to incollection
JOSS system doesn't use the 'note' field, so let's move back to
incollection to use the field editor.
