+++ joss/paper.bib
@@ -1,11 +1,11 @@
 
 
-@manual{WOD2018,
-  title = {{World Ocean Database} 2018},
+@incollection{WOD2018,
+  title = {{W}orld {O}cean {D}atabase 2018},
   author = {Boyer, T. P. and O. K. Baranova and  C. Coleman and  H. E. Garcia and  A. Grodsky and  R. A. Locarnini and  A. V. Mishonov and  C. R. Paver and  J. R. Reagan and  D. Seidov and  I. V. Smolyar and  K. Weathers and  M. M. Zweng},
   year = {2019},
-  note = {A. V. Mishonov, Technical Editor},
-  series = {{NOAA} {Atlas} {NESDIS} 87},
+  editor = {A. V. Mishonov},
+  series = {{NOAA} {A}tlas {NESDIS} 87},
 }
 
 @manual{Argo:2015,
