### 2020-03-31 #2063 @castelao
@kthyng I spoke with Tim Boyer today, the first author of that publication, and he said that on his side there is no problem with using 'Editor' instead of 'Technical Editor'.
