### 2020-04-07 #2063 @castelao
@kthyng I just updated the DOI record itself, thanks for noticing that.

The current version is 0.21.3
